+++ lib/bitcrowd_ecto/changeset.ex
@@ -425,4 +425,176 @@ defmodule BitcrowdEcto.Changeset do
       end
     end
   end
+
+  @type auto_cast_option :: {:action, atom}
+
+  @doc """
+  Introspects a schema and casts all defined fields from a params map.
+
+  - Accepts a schema module or structs or changesets.
+  - Can deal with embeds.
+
+  ## Options
+
+  - `required` list of required field names
+  - `optional` list of optional field names (= inverse set is required)
+
+  `required` and `optional` options must not be present at the same time.
+
+  ## Examples
+
+      iex> changeset = auto_cast(TestEmbeddedSchema, %{some_field: 4})
+      iex> changeset.valid?
+      true
+
+      iex> changeset = auto_cast(%TestEmbeddedSchema{}, %{some_field: 4})
+      iex> changeset.valid?
+      true
+
+      iex> changeset = auto_cast(change(%TestEmbeddedSchema{}), %{some_field: 4})
+      iex> changeset.valid?
+      true
+
+      iex> changeset = auto_cast(TestEmbeddedSchema, %{}, required: [:some_field])
+      iex> changeset.errors
+      [some_field: {"can't be blank", [validation: :required]}]
+
+      iex> changeset = auto_cast(TestEmbeddedSchema, %{}, optional: [:some_other_field])
+      iex> changeset.errors
+      [some_field: {"can't be blank", [validation: :required]}]
+
+  """
+  @doc since: "0.17.0"
+  @spec auto_cast(module | struct, map) :: Ecto.Changeset.t()
+  @spec auto_cast(module | struct, map, [auto_cast_option]) :: Ecto.Changeset.t()
+  def auto_cast(schema_or_struct_or_changeset, params, opts \\ [])
+
+  def auto_cast(%Ecto.Changeset{} = changeset, params, opts) do
+    do_auto_cast(changeset.data.__struct__, changeset, params, opts)
+  end
+
+  def auto_cast(schema, params, opts) when is_atom(schema) do
+    do_auto_cast(schema, struct!(schema), params, opts)
+  end
+
+  def auto_cast(struct, params, opts) when is_struct(struct) do
+    do_auto_cast(struct.__struct__, struct, params, opts)
+  end
+
+  defp do_auto_cast(schema, struct_or_changeset, params, opts) do
+    required = required_fields(schema, opts)
+    grouped = grouped_fields(schema)
+
+    struct_or_changeset
+    |> cast_scalars(params, grouped.scalars, required)
+    |> cast_embeds(grouped.embeds, required)
+  end
+
+  defp required_fields(schema, opts) do
+    required = Keyword.get(opts, :required)
+    optional = Keyword.get(opts, :optional)
+
+    cond do
+      required && optional ->
+        raise ArgumentError, ":required and :optional options are mutually exclusive"
+
+      required ->
+        required
+
+      optional ->
+        schema.__schema__(:fields) -- optional
+
+      true ->
+        []
+    end
+  end
+
+  defp grouped_fields(schema) do
+    :fields
+    |> schema.__schema__()
+    |> Enum.group_by(fn field ->
+      case schema.__schema__(:type, field) do
+        {:parameterized, Ecto.Embedded, _} ->
+          :embeds
+
+        # Simplification, can be extended as needed.
+        _other ->
+          :scalars
+      end
+    end)
+    |> Map.put_new(:embeds, [])
+    |> Map.put_new(:scalars, [])
+  end
+
+  defp cast_scalars(schema_struct, params, scalars, required) do
+    required = scalars -- scalars -- required
+
+    schema_struct
+    |> Ecto.Changeset.cast(params, scalars)
+    |> Ecto.Changeset.validate_required(required)
+  end
+
+  defp cast_embeds(changeset, embeds, required) do
+    Enum.reduce(embeds, changeset, fn embed, cs ->
+      Ecto.Changeset.cast_embed(cs, embed, required: embed in required)
+    end)
+  end
+
+  @type action_option :: {:action, :insert | :update}
+
+  @doc """
+  Convenience function calling `auto_cast/3` and `apply_action/2` in one step.
+
+  ## Options
+
+  - `action` the action to apply to the changeset (`:insert` or `:update`, defaults to `:insert`)
+  - all options from `auto_cast/3`
+
+  ## Example
+
+      iex> auto_cast_and_apply(TestEmbeddedSchema, %{some_field: 4})
+      {:ok, %TestEmbeddedSchema{some_field: 4}}
+  """
+  @doc since: "0.17.0"
+  @spec auto_cast_and_apply(module | struct, map) ::
+          {:ok, Ecto.Schema.t()} | {:error, Ecto.Changeset.t()}
+  @spec auto_cast_and_apply(module | struct, map, [auto_cast_option | action_option]) ::
+          {:ok, Ecto.Schema.t()} | {:error, Ecto.Changeset.t()}
+  def auto_cast_and_apply(schema_or_struct_or_changeset, params, opts \\ []) do
+    do_auto_cast_and_apply(
+      schema_or_struct_or_changeset,
+      params,
+      &Ecto.Changeset.apply_action/2,
+      opts
+    )
+  end
+
+  @doc """
+  Explosive variant of `auto_cast_and_apply/3`.
+
+  ## Example
+
+      iex> auto_cast_and_apply!(TestEmbeddedSchema, %{some_field: 4})
+      %TestEmbeddedSchema{some_field: 4}
+  """
+  @doc since: "0.17.0"
+  @spec auto_cast_and_apply!(module | struct, map) :: Ecto.Schema.t() | no_return
+  @spec auto_cast_and_apply!(module | struct, map, [auto_cast_option | action_option]) ::
+          Ecto.Schema.t() | no_return
+  def auto_cast_and_apply!(schema_or_struct_or_changeset, params, opts \\ []) do
+    do_auto_cast_and_apply(
+      schema_or_struct_or_changeset,
+      params,
+      &Ecto.Changeset.apply_action!/2,
+      opts
+    )
+  end
+
+  defp do_auto_cast_and_apply(schema_or_struct_or_changeset, params, apply_action, opts) do
+    {action, opts} = Keyword.pop(opts, :action, :insert)
+
+    schema_or_struct_or_changeset
+    |> auto_cast(params, opts)
+    |> apply_action.(action)
+  end
 end